--- .idea/checkstyle-idea.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CheckStyle-IDEA" serialisationVersion="2">
-    <checkstyleVersion>10.13.0</checkstyleVersion>
-    <scanScope>JavaOnly</scanScope>
-    <copyLibs>true</copyLibs>
-    <option name="thirdPartyClasspath" />
-    <option name="activeLocationIds">
-      <option value="06ebdd06-7c6f-4106-a14c-d7a089398476" />
-    </option>
-    <option name="locations">
-      <list>
-        <ConfigurationLocation id="bundled-sun-checks" type="BUNDLED" scope="All" description="Sun Checks">(bundled)</ConfigurationLocation>
-        <ConfigurationLocation id="bundled-google-checks" type="BUNDLED" scope="All" description="Google Checks">(bundled)</ConfigurationLocation>
-        <ConfigurationLocation id="06ebdd06-7c6f-4106-a14c-d7a089398476" type="LOCAL_FILE" scope="All" description="SpringCheckstyle">$PROJECT_DIR$/checkstyle.xml</ConfigurationLocation>
-      </list>
-    </option>
-  </component>
-</project>
