+++ bot/src/main/java/edu/java/bot/CommandPool.java
@@ -4,20 +4,35 @@ import com.pengrad.telegrambot.model.BotCommand;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
 import edu.java.bot.commands.Command;
+import edu.java.bot.commands.HelpCommand;
+import edu.java.bot.commands.ListCommand;
+import edu.java.bot.commands.StartCommand;
+import edu.java.bot.commands.TrackCommand;
+import edu.java.bot.commands.UntrackCommand;
+import java.util.ArrayList;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 
-@Component
 public class CommandPool {
 
     private static final String NO_SUPPORT_MESSAGE = "Команда не поддерживается или написана неправильно!";
 
-    private final List<? extends Command> commands;
+    private final List<Command> commands;
 
-    @Autowired
-    public CommandPool(List<? extends Command> commands) {
+    public CommandPool(List<Command> commands) {
         this.commands = commands;
+    }
+
+    public static CommandPool standardPool(LinkStorage linkStorage) {
+        List<Command> tmpCommands = new ArrayList<>(List.of(
+            new StartCommand(linkStorage),
+            new TrackCommand(linkStorage),
+            new UntrackCommand(linkStorage),
+            new ListCommand(linkStorage),
+            new HelpCommand(List.of())
+        ));
+        ((HelpCommand) tmpCommands.getLast()).setCommands(tmpCommands);
+
+        return new CommandPool(tmpCommands);
     }
 
     public List<BotCommand> getBotCommands() {
