+++ bot/src/main/java/edu/java/bot/LinkStorage.java
@@ -1,4 +1,4 @@
-package edu.java.bot.storage;
+package edu.java.bot;
 
 import java.util.List;
