+++ bot/src/test/java/edu/java/bot/CommandPoolTest.java
@@ -5,19 +5,11 @@ import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.model.request.ParseMode;
 import com.pengrad.telegrambot.request.SendMessage;
-import edu.java.bot.commands.Command;
-import edu.java.bot.commands.HelpCommand;
-import edu.java.bot.commands.ListCommand;
-import edu.java.bot.commands.StartCommand;
-import edu.java.bot.commands.TrackCommand;
-import edu.java.bot.commands.UntrackCommand;
-import edu.java.bot.storage.LinkStorage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mockito;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -26,21 +18,8 @@ import static org.mockito.ArgumentMatchers.any;
 
 class CommandPoolTest {
 
-    public static CommandPool standardPool(LinkStorage linkStorage) {
-        List<Command> tmpCommands = new ArrayList<>(List.of(
-            new StartCommand(linkStorage),
-            new TrackCommand(linkStorage),
-            new UntrackCommand(linkStorage),
-            new ListCommand(linkStorage),
-            new HelpCommand(List.of())
-        ));
-        HelpCommand helpCommand = (HelpCommand) tmpCommands.getLast();
-        helpCommand.setCommands(tmpCommands.stream()
-            .map(Command::botCommand)
-            .toList());
-
-        return new CommandPool(tmpCommands);
-    }
+    private final static CommandPool commandPool =
+        CommandPool.standardPool(generateStorageWithLinks(Collections.emptyList()));
 
     private static Update generateUpdateWithIdAndContent(String content, Long id) {
         Update update = Mockito.mock(Update.class);
@@ -68,7 +47,6 @@ class CommandPoolTest {
     @ValueSource(strings = {"", " ", "track", "help", "/end", "/", "/ "})
     @DisplayName("Сообщения не содержат поддерживаемых команд")
     void process_shouldReturnSpecialMessageIfMessageHasNoCommand(String content) {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(Collections.emptyList()));
         Update update = generateUpdateWithIdAndContent(content, 0L);
 
         SendMessage result = commandPool.process(update);
@@ -82,7 +60,6 @@ class CommandPoolTest {
     @Test
     @DisplayName("/list с пустым списком")
     void process_shouldReturnSpecialMessageForListIfLinkListEmpty() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(Collections.emptyList()));
         Update update = generateUpdateWithIdAndContent("/list", 0L);
 
         SendMessage result = commandPool.process(update);
@@ -98,7 +75,7 @@ class CommandPoolTest {
     @Test
     @DisplayName("/list с непустым списком")
     void process_shouldReturnMessageWithLinkListForNotEmptyList() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(List.of(
+        CommandPool commandPool = CommandPool.standardPool(generateStorageWithLinks(List.of(
             "first",
             "second"
         )));
@@ -117,7 +94,6 @@ class CommandPoolTest {
     @Test
     @DisplayName("/help")
     void process_shouldReturnCommandListForHelp() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(Collections.emptyList()));
         Update update = generateUpdateWithIdAndContent("/help", 0L);
 
         SendMessage result = commandPool.process(update);
@@ -135,7 +111,7 @@ class CommandPoolTest {
     @Test
     @DisplayName("/start для незарегистрированного пользователя")
     void process_shouldReturnDefaultStartMessage() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(null));
+        CommandPool commandPool = CommandPool.standardPool(generateStorageWithLinks(null));
         Update update = generateUpdateWithIdAndContent("/start", 0L);
         Mockito.when(update.message().chat().firstName()).thenReturn("name");
 
@@ -150,7 +126,6 @@ class CommandPoolTest {
     @Test
     @DisplayName("/start для зарегистрированного пользователя")
     void process_shouldReturnSpecialStartMessageForRegisteredUser() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(Collections.emptyList()));
         Update update = generateUpdateWithIdAndContent("/start", 0L);
 
         SendMessage result = commandPool.process(update);
@@ -164,7 +139,6 @@ class CommandPoolTest {
     @Test
     @DisplayName("/track для корректной ссылки")
     void process_shouldReturnDefaultMessageForTrackingValidLink() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(Collections.emptyList()));
         Update update = generateUpdateWithIdAndContent("/track https://www.youtube.com/watch?v=4i2ifGa5n0o", 0L);
 
         SendMessage result = commandPool.process(update);
@@ -180,7 +154,6 @@ class CommandPoolTest {
     @Test
     @DisplayName("/track для некорректной ссылки")
     void process_shouldReturnSpecialMessageForTrackingNotValidLink() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(Collections.emptyList()));
         Update update = generateUpdateWithIdAndContent("/track https://www.youtube.o/watch?v=4i2ifGa5n0o", 0L);
 
         SendMessage result = commandPool.process(update);
@@ -196,7 +169,7 @@ class CommandPoolTest {
     @DisplayName("/untrack для ссылки в списке")
     void process_shouldReturnDefaultMessageForUntrackingTrackingLink() {
         CommandPool commandPool =
-            standardPool(generateStorageWithLinks(List.of("https://www.youtube.com/watch?v=4i2ifGa5n0o")));
+            CommandPool.standardPool(generateStorageWithLinks(List.of("https://www.youtube.com/watch?v=4i2ifGa5n0o")));
         Update update = generateUpdateWithIdAndContent("/untrack https://www.youtube.com/watch?v=4i2ifGa5n0o", 0L);
 
         SendMessage result = commandPool.process(update);
@@ -212,7 +185,6 @@ class CommandPoolTest {
     @Test
     @DisplayName("/untrack для ссылки не списке")
     void process_shouldReturnSpecialMessageForUntrackingNotTrackingLink() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(Collections.emptyList()));
         Update update = generateUpdateWithIdAndContent("/untrack https://www.youtube.com/watch?v=4i2ifGa5n0o", 0L);
 
         SendMessage result = commandPool.process(update);
@@ -228,7 +200,6 @@ class CommandPoolTest {
     @Test
     @DisplayName("/untrack для некорректной ссылки")
     void process_shouldReturnSpecialMessageForUntrackingNotValidLink() {
-        CommandPool commandPool = standardPool(generateStorageWithLinks(Collections.emptyList()));
         Update update = generateUpdateWithIdAndContent("/untrack https://www.youtube.om/watch?v=4i2ifGa5n0o", 0L);
 
         SendMessage result = commandPool.process(update);
